refactor: split mappers into logical modules
