+++ src/parser/mappers/index.ts
@@ -1,3 +1,4 @@
-export * from "./noteMeasureMappers";
+export * from "./noteMappers";
+export * from "./measureMappers";
 export * from "./creditMappers";
 export * from "./defaultsMappers";
