+++ src/parser/mappers/utils.ts
@@ -44,3 +44,19 @@ export const parseOptionalNumberAttribute = (
   const num = parseInt(value, 10);
   return isNaN(num) ? undefined : num;
 };
+
+export const parseOptionalInt = (
+  value: string | null | undefined,
+): number | undefined => {
+  if (value === null || value === undefined || value.trim() === "") return undefined;
+  const num = parseInt(value, 10);
+  return isNaN(num) ? undefined : num;
+};
+
+export const parseOptionalFloat = (
+  value: string | null | undefined,
+): number | undefined => {
+  if (value === null || value === undefined || value.trim() === "") return undefined;
+  const num = parseFloat(value);
+  return isNaN(num) ? undefined : num;
+};
